--- a/benchmarks/.gitignore
+++ b/benchmarks/.gitignore
@@ -1,2 +1,5 @@
 # clinic.js
 **clinic**
+
+# generated test files
+fixtures