--- a/benchmarks/package.json
+++ b/benchmarks/package.json
@@ -4,8 +4,11 @@
   "description": "",
   "main": "bench.js",
   "scripts": {
+    "setup": "node util/create-files",
     "pull-mplex": "node bench.js --lib=pull-mplex",
-    "libp2p-mplex": "node bench.js --lib=libp2p-mplex"
+    "libp2p-mplex": "node bench.js --lib=libp2p-mplex",
+    "lg-pull-mplex": "node large-file.js --lib=pull-mplex",
+    "lg-libp2p-mplex": "node large-file.js --lib=libp2p-mplex"
   },
   "keywords": [],
   "author": "",
@@ -15,6 +18,7 @@
     "fastparallel": "^2.3.0",
     "libp2p-mplex": "~0.8.4",
     "minimist": "^1.2.0",
+    "pull-file": "^1.1.0",
     "pull-mplex": "../",
     "pull-stream": "^3.6.9",
     "stream-to-pull-stream": "^1.7.2"